+++ crates/nixci/src/command/mod.rs
@@ -1,3 +1,3 @@
-pub mod build;
 pub mod core;
-pub mod gh;
+pub mod gh_matrix;
+pub mod run;